requirements.txt
requests==2.21.0
bitcoin_utils==0.3.2
python-dotenv==0.13.0
